Move JsonSkinSerializer
